Modularize and enhance the headless test script.
This refactors the test-script `run-testsuite`, so it isn't one big file,
but uses two newly created modules: `strUtil` and `tawWrapper`.

* `strUtil` exports methods for padding strings.
* `tawWrapper` exports methods for finding test-files, plus finding and
  executing Test.Anotherway functions.

Also included in this commit is a more robust way of detecting the base path
of the testsuite. This script should now be callable from everywhere, not only
from the root of the repository.
